Update docker-compose instructions
Figured we should stop directing users to an old docker-compose binary.

In addition, this switches from the -p flag to COMPOSE_PROJECT_NAME to
ensure that environment variable is set. -p behaves the same but does
not set the variable. See the following link for more.

https://github.com/docker/compose/issues/3431#issuecomment-221291167

Finally, I added the -v flag to docker-compose down, which ensures any
automatic volumes are removed when the cluster is stopped.