Do not advise TLS off for any logging service
This encourages people to send their query texts and the like as
plaintext over the Internet!  A quizzical customer pointed this out,
for a service where the only thing necessary to do was to change the
port and check the box, Logentries.

I since have doubled-back and verified:

* LogDNA
* Logentries / Rapid7 insightOps (acquired name)
* Loggly
* Papertrail

Two of the other providers require their certificate be installed;
thus, this patch must come after
https://github.com/citusdata/cloudplane/pull/957 is deployed.
Logentries uses GoDaddy and does not give cert bundle guidence,
Papertrail provides a cert bundle, but it is also a major vendor and
handled by the Mozilla bundle.

I suppose in principle it might be nice if the customer could define
the CA expected, or we only installed necessary certs for the vendor
selected, but, alas: complexity, and the current situation is much
worse.

Unfortunately, so far, we have been unable to turn up any rationale
for why we were so lackadaisical about TLS in our guidance: it simply
always was and passed without detectable remark in pull requests,
commit texts, comments, and emails.

The two times this could have been detected were when the CloudPlane
stuff was committed -- I was the reviewer, and could have asked why
TLS was optional -- and when the manual was updated to include the
three TLS-free provider directives.  Unlike obliged SOC2 procedures
nowadays, CloudCon's commit of this was unilateral. It was in Aug
2016, and the first customers were Apr 2016 or so.

Some hyperlinks from when Mistakes Were Made:

https://github.com/citusdata/citus_docs/pull/413
https://github.com/citusdata/cloudcon/pull/45
https://github.com/citusdata/cloudplane/pull/145