Update "Adding a Coordinator" to reflect the latest behaviour. (#1096)
Adding a Coordinator section presents some stale information. The second paragraph, starting with "However ..." 
should either be removed or updated. 

Removing does not make sense since the section will not have anything to do with "Adding a Coordinator" heading.  

I suggest removing the "Adding a Coordinator" altogether or update in info as in this PR.